Move NavBar to one single file, only to be called by the Router
